Flower/bcfl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bcfl"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.19.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "juanluis"

[tool.flwr.app.components]
serverapp = "bcfl.server_app:app"
clientapp = "bcfl.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 10
fraction-fit = 0.29412
local-epochs = 5

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 17
